# documentation/openapi.yaml
openapi: 3.1.0
info:
  title: Context Board MVP API
  version: 1.0.0
  summary: REST API for a single-user AI-first visual whiteboard MVP
  description: |
    Context Board is an AI-first visual workspace with an infinite canvas and a board-aware chat agent.

    MVP constraints:
    - single-user only
    - no authentication
    - one board = one chat thread
    - backend is source of truth
    - all durable mutations are logged
    - agent edits use suggest/apply flow
servers:
  - url: /api
...
paths:
  /boards:
    get:
      tags: [boards]
      summary: List boards
      operationId: listBoards
      responses:
        '200':
          description: Board list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardListResponse'
    post:
      tags: [boards]
      summary: Create board
      operationId: createBoard
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBoardRequest'
      responses:
        '201':
          description: Board created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBoardResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /boards/{boardId}:
    get:
      tags: [boards]
      summary: Get board
      operationId: getBoard
      parameters:
        - $ref: '#/components/parameters/BoardId'
      responses:
        '200':
          description: Board metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetBoardResponse'
        '404':
          $ref: '#/components/responses/BoardNotFound'
    patch:
      tags: [boards]
      summary: Update board
      operationId: updateBoard
      parameters:
        - $ref: '#/components/parameters/BoardId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateBoardRequest'
      responses:
        '200':
          description: Board updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetBoardResponse'
        '404':
          $ref: '#/components/responses/BoardNotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
    delete:
      tags: [boards]
      summary: Soft-delete board
      operationId: deleteBoard
      parameters:
        - $ref: '#/components/parameters/BoardId'
      responses:
        '200':
          description: Board deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteBoardResponse'
        '404':
          $ref: '#/components/responses/BoardNotFound'
...
components:
  parameters:
    BoardId:
      name: boardId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 255
  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            error:
              code: VALIDATION_ERROR
              message: Invalid request
              details: {}
    BoardNotFound:
      description: Board not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            error:
              code: BOARD_NOT_FOUND
              message: Board not found
              details: {}
...
    UnsupportedMediaType:
      description: Unsupported media type
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            data: null
            error:
              code: UNSUPPORTED_MEDIA_TYPE
              message: Unsupported media type
              details: {}
...
  schemas:
...
    ErrorObject:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      type: object
      required: [data, error]
      properties:
        data:
          type: 'null'
        error:
          $ref: '#/components/schemas/ErrorObject'
    SuccessEnvelopeBase:
      type: object
      required: [data, error]
      properties:
        error:
          type: 'null'
    BoardStatus:
      type: string
      enum: [active, archived, deleted]
...
    AgentMode:
      type: string
      enum: [suggest, apply]
...
    ViewportState:
      type: object
      additionalProperties: false
      required: [x, y, zoom]
      properties:
        x:
          type: number
        y:
          type: number
        zoom:
          type: number
          exclusiveMinimum: 0
    BoardSettings:
      type: object
      additionalProperties: true
      properties:
        gridEnabled:
          type: boolean
        snapToGrid:
          type: boolean
        agentEditMode:
          $ref: '#/components/schemas/AgentMode'
    BoardSummary:
      type: object
      additionalProperties: true
    Board:
      type: object
      required:
        - id
        - title
        - description
        - status
        - viewportState
        - settings
        - summary
        - revision
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          nullable: true
          maxLength: 10000
        status:
          $ref: '#/components/schemas/BoardStatus'
        viewportState:
          $ref: '#/components/schemas/ViewportState'
        settings:
          $ref: '#/components/schemas/BoardSettings'
        summary:
          $ref: '#/components/schemas/BoardSummary'
        revision:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ChatThread:
      type: object
      required: [id, boardId, metadata, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        boardId:
          type: string
          format: uuid
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    CreateBoardRequest:
      type: object
      required: [title]
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 10000
    UpdateBoardRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          nullable: true
          maxLength: 10000
        viewportState:
          $ref: '#/components/schemas/ViewportState'
        settings:
          $ref: '#/components/schemas/BoardSettings'
        summary:
          $ref: '#/components/schemas/BoardSummary'
        status:
          type: string
          description: >
            Status transition via PATCH. Only 'archived' is accepted
            (active → archived). All other transitions return 422.
...
    BoardListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelopeBase'
        - type: object
          properties:
            data:
              type: object
              required: [boards]
              properties:
                boards:
                  type: array
                  items:
                    $ref: '#/components/schemas/Board'
    CreateBoardResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelopeBase'
        - type: object
          properties:
            data:
              type: object
              required: [board, chatThread]
              properties:
                board:
                  $ref: '#/components/schemas/Board'
                chatThread:
                  $ref: '#/components/schemas/ChatThread'
    GetBoardResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelopeBase'
        - type: object
          properties:
            data:
              type: object
              required: [board]
              properties:
                board:
                  $ref: '#/components/schemas/Board'
    DeleteBoardResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelopeBase'
        - type: object
          properties:
            data:
              type: object
              required: [success, boardId]
              properties:
                success:
                  type: boolean
                boardId:
                  type: string
                  format: uuid
...